Created logic and base framework for switching and user input
The idea here is that when the user kicks off the script there needs to be
conditional logic to decide what will need to happen.  In the beginning it
is just going to have the ability to discover the configuration, but
eventually this will be able to transfer logic to failing over,
disabling/enabling log shipping, etc.